% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_resistance_indicator.R
\name{compute_resistance_indicator}
\alias{compute_resistance_indicator}
\title{Computes the resistance indicator based on raw AMR surveillance data}
\usage{
compute_resistance_indicator(raw_data)
}
\arguments{
\item{raw_data}{A list of data frames.}
}
\value{
A data frame.
}
\description{
The indicator is the resistance rate from the WHO EURO AMR dashboard with a
few variables renamed for coherence across indicators.
}
