% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_UHC_service_coverage_indicator.R
\name{compute_UHC_service_coverage_indicator}
\alias{compute_UHC_service_coverage_indicator}
\title{Computes the UHC service coverage indicator based on raw AMR surveillance data}
\usage{
compute_UHC_service_coverage_indicator(raw_data)
}
\arguments{
\item{raw_data}{A list of data frames.}
}
\value{
A data frame.
}
\description{
The indicator is the GHO indicator UHC_SCI_INFECT with a few variables
renamed for coherence across indicators.
}
