% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vaccination_rate_indicator.R
\name{compute_vaccination_rate_indicator}
\alias{compute_vaccination_rate_indicator}
\title{Computes the vaccination rate indicator based on raw AMR surveillance data}
\usage{
compute_vaccination_rate_indicator(raw_data)
}
\arguments{
\item{raw_data}{A list of data frames.}
}
\value{
A data frame.
}
\description{
The indicator is defined as the maximum of the GHO indicators PCV3 and
WHS4_129, with a few variables renamed for coherence across indicators.
}
