% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_time.R
\name{draw_line_plot_time}
\alias{draw_line_plot_time}
\title{Draws a line plot of the indicator against time}
\usage{
draw_line_plot_time(indicator, plot_title)
}
\arguments{
\item{indicator}{A data frame. The indicator to be plotted: the variable plotted has to be named \code{value}, the time variable is \code{year}.}

\item{plot_title}{A string. The title of the plot.}
}
\value{
A ggplot object.
}
\description{
Draws a line plot of the indicator against time
}
