% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_map}
\alias{draw_map}
\title{Draws a map of the indicator chosen for the last year data are available.}
\usage{
draw_map(indicator, geo_shapes, plot_title)
}
\arguments{
\item{indicator}{A data frame. The indicator to be plotted: the variable plotted has to be named \code{value}, the time variable is \code{year}, the regions are \code{region_code_iso3}.}

\item{plot_title}{A string. The title of the plot.}
}
\value{
A ggplot object.
}
\description{
Draws a map of the indicator chosen for the last year data are available.
}
