% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw AMR surveillance data}
\usage{
load_raw_data(
  data_directories,
  data_source_list,
  API_accessible,
  download_from_API,
  gho_api
)
}
\arguments{
\item{data_directories}{List of directories where the data to load from files are located.}

\item{data_source_list}{List of data sources, currently "EURO", "GHO", or "IHME"}

\item{API_accessible}{Boolean: whether a source has an API from which one can download the data. (If not, it means one has to download data manually.)}

\item{download_from_API}{Whether to download data from the API's. Applies to all sources with accessible API's.}

\item{gho_api}{GHO API address.}
}
\value{
A list of data frames.
}
\description{
\code{load_raw_data()} loads raw AMR surveillance data from files or download them. Currently only GHO has an accessible API.
}
\examples{
\dontrun{
data_directories <- c(
  EURO = "../AMR-raw-data/EURO",
  GHO = "../AMR-raw-data/GHO",
  IHME = "../AMR-raw-data/IHME"
)
data_source_list <- c("EURO", "GHO", "IHME")
API_accessible <- c(EURO = FALSE, GHO = TRUE, IHME = FALSE)
download_from_API <- FALSE
gho_api <- "https://ghoapi.azureedge.net/api/"
raw_AMR_surv_data <- load_raw_data(data_directories, data_source_list,
  API_accessible, download_from_API, gho_api)
}
}
